@@ -3,7 +3,7 @@ requires = ["setuptools", "wheel"]
 
 [tool.ruff]
 src = ["src"]
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 select = [
