@@ -1,10 +1,9 @@
 from datetime import datetime, tzinfo
-from typing import Optional
 
 
 def is_aware(dt: datetime) -> bool:
     return dt.tzinfo is not None and dt.tzinfo.utcoffset(dt) is not None
 
 
-def to_tz_unaware(dt: datetime, tz: Optional[tzinfo]) -> datetime:
+def to_tz_unaware(dt: datetime, tz: tzinfo | None) -> datetime:
     return dt.astimezone(tz).replace(tzinfo=None)
